requirements.txt
Django==5.2.7
django-jazzmin
psycopg
django-filter
python-dotenv
django-solo
django-crum
django-extensions
Faker
djangorestframework
djangorestframework-simplejwt
django-jazzmin
python-dotenv
drf-spectacular
drf-spectacular-sidecar
django-cors-headers
psycopg2-binary; sys_platform == 'win32'
gunicorn
cloudinary
Pillow
ruff


